services/payment_service/requirements.txt
Django==5.2.6
djangorestframework==3.15.2
psycopg2-binary==2.9.9
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
stripe==9.12.0          # Stripe SDK (можно заменить на другой процессор)
pytest==8.4.2
pytest-django==4.11.1